Create Kaggle_Credit_Score
This project encompasses the development of a sophisticated credit scoring model aimed at advancing the current state-of-the-art. The primary objective is to predict the likelihood of an individual encountering financial distress within the upcoming two years. The foundation of this model is rooted in addressing the challenge presented by Kaggle.com, known as "Give Me Some Credit."